+++ src/applications/health-care-questionnaire/questionnaire/tests/unit/reducer/reducer.loading.appointment.data.unit.spec.js
@@ -7,7 +7,7 @@ import {
   questionnaireAppointmentLoading,
 } from '../../../actions';
 
-import testData from '../../../api/appointment-data.json';
+import testData from '../../../../shared/api/appointment-data.json';
 
 describe('health care questionnaire -- questionnaire reducer --', () => {
   it('should set loading to true', () => {
